+++ source/Public/Invoke-Git.ps1
@@ -5,11 +5,16 @@
     .DESCRIPTION
         Invokes a git command with command line arguments using System.Diagnostics.Process.
 
+        Throws an error when git ExitCode -ne 0 and -PassThru switch -eq $false (or omitted).
+
     .PARAMETER WorkingDirectory
         The path to the git working directory.
 
     .PARAMETER Timeout
         Milliseconds to wait for process to exit.
+
+    .PARAMETER PassThru
+        Switch parameter when enabled will return result object of running git command.
 
     .PARAMETER Arguments
         The arguments to pass to the Git executable.
@@ -20,7 +25,7 @@
         Invokes the Git executable to clone the specified repository to the working directory.
 
     .EXAMPLE
-        Invoke-Git -WorkingDirectory 'C:\SomeDirectory' -Arguments @( 'status' ) -TimeOut 10000
+        Invoke-Git -WorkingDirectory 'C:\SomeDirectory' -Arguments @( 'status' ) -TimeOut 10000 -PassThru
 
         Invokes the Git executable to return the status while having a 10000 millisecond timeout.
 #>
@@ -35,30 +40,24 @@ function Invoke-Git
         [System.String]
         $WorkingDirectory,
 
-        [Parameter(Mandatory = $false)]
+        [Parameter()]
         [System.Int32]
         $TimeOut = 120000,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $PassThru,
 
         [Parameter(ValueFromRemainingArguments = $true)]
         [System.String[]]
         $Arguments
     )
 
-    $returnValue = @{
-        'ExitCode'       = -1
-        'StandardOutput' = ''
-        'StandardError'  = ''
+    $gitResult = @{
+        'ExitCode'         = -1
+        'StandardOutput'   = ''
+        'StandardError'    = ''
     }
-
-    $argumentsJoined = $Arguments -join ' '
-
-    # Trying to remove any access token from the debug output.
-    if ($argumentsJoined -match ':[\d|a-f].*@')
-    {
-        $argumentsJoined = $argumentsJoined -replace ':[\d|a-f].*@', ':RedactedToken@'
-    }
-
-    Write-Debug -Message ($localizedData.InvokingGitMessage -f $argumentsJoined)
 
     try
     {
@@ -76,13 +75,9 @@ function Invoke-Git
         {
             if ($process.WaitForExit($TimeOut) -eq $true)
             {
-                $returnValue.ExitCode = $process.ExitCode
-                $returnValue.StandardOutput = $process.StandardOutput.ReadToEnd()
-                $returnValue.StandardError = $process.StandardError.ReadToEnd()
-
-                # Remove all new lines at end of string.
-                $returnValue.StandardOutput = $returnValue.StandardOutput -replace '[\r?\n]+$'
-                $returnValue.StandardError = $returnValue.StandardError -replace '[\r?\n]+$'
+                $gitResult.ExitCode = $process.ExitCode
+                $gitResult.StandardOutput = $process.StandardOutput.ReadToEnd()
+                $gitResult.StandardError = $process.StandardError.ReadToEnd()
             }
         }
     }
@@ -96,11 +91,33 @@ function Invoke-Git
         {
             $process.Dispose()
         }
+
+        if ($VerbosePreference -ne 'SilentlyContinue' -or `
+            $DebugPreference -ne 'SilentlyContinue' -or `
+            $PSBoundParameters['Verbose'] -eq $true -or `
+            $PSBoundParameters['Debug'] -eq $true)
+        {
+            $outGitResultSplat = $gitResult.Clone()
+
+            $outGitResultSplat.Add('Command', $Arguments)
+            $outGitResultSplat.Add('WorkingDirectory', $WorkingDirectory)
+
+            Out-GitResult @outGitResultSplat
+        }
+
+        if ($gitResult.ExitCode -ne 0 -and $PassThru -eq $false)
+        {
+            $throwMessage = "$($script:localizedData.InvokeGitCommandDebug -f $(Hide-GitToken -Command $Arguments))`n" +`
+                            "$($script:localizedData.InvokeGitExitCodeMessage -f $gitResult.ExitCode)`n" +`
+                            "$($script:localizedData.InvokeGitStandardOutputMessage -f $gitResult.StandardOutput)`n" +`
+                            "$($script:localizedData.InvokeGitStandardErrorMessage -f $gitResult.StandardError)`n"
+
+            throw $throwMessage
+        }
     }
 
-    Write-Debug -Message ('{0}: {1}' -f $MyInvocation.MyCommand.Name, ($localizedData.InvokeGitExitCodeMessage -f $returnValue.ExitCode))
-    Write-Debug -Message ('{0}: {1}' -f $MyInvocation.MyCommand.Name, ($localizedData.InvokeGitStandardOutputMessage -f $returnValue.StandardOutput))
-    Write-Debug -Message ('{0}: {1}' -f $MyInvocation.MyCommand.Name, ($localizedData.InvokeGitStandardErrorMessage -f $returnValue.StandardError))
-
-    return $returnValue
+    if ($PSBoundParameters['PassThru'] -eq $true)
+    {
+        return $gitResult
+    }
 }