+++ tests/unit/public/Invoke-Git.Tests.ps1
@@ -24,100 +24,197 @@ InModuleScope $script:moduleName {
             $mockProcess = New-MockObject -Type System.Diagnostics.Process
             $mockProcess | Add-Member -MemberType ScriptMethod -Name 'Start' -Value { $true } -Force
             $mockProcess | Add-Member -MemberType ScriptMethod -Name 'WaitForExit' -Value { $true } -Force
-            $mockProcess | Add-Member -MemberType ScriptProperty -Name ExitCode -Value { 0 } -Force
             $mockProcess | Add-Member -MemberType ScriptProperty -Name WorkingDirectory -Value { '' } -Force
 
+            $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardOutput' -Value {
+                New-Object -TypeName 'Object' | `
+                    Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Output Message' } -PassThru -Force
+            } -Force
+
+            $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardError' -Value {
+                New-Object -TypeName 'Object' | `
+                    Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Error Message' } -PassThru -Force
+            } -Force
+
             Mock -CommandName New-Object -MockWith { return $mockProcess } -ParameterFilter { $TypeName -eq 'System.Diagnostics.Process' } -Verifiable
+            Mock -CommandName Out-GitResult
         }
 
-        Context 'When calling Invoke-Git' {
+        Context 'When git ExitCode -eq 0' {
             BeforeAll {
-                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardOutput' -Value {
-                    New-Object -TypeName 'Object' | `
-                        Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Output Message 0' } -PassThru -Force
-                } -Force
-
-                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardError' -Value {
-                    New-Object -TypeName 'Object' | `
-                        Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Error Message 0' } -PassThru -Force
-                } -Force
+                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'ExitCode' -Value { 0 } -Force
             }
-            It 'Should complete with ExitCode=0' {
 
-                $result = Invoke-Git -WorkingDirectory $TestDrive `
-                                -Arguments @( 'config', '--local', 'user.email', 'user@host.com' )
+            It 'Should not throw, return result with -PassThru' {
+                $result = Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -PassThru
 
                 $result.ExitCode | Should -BeExactly 0
 
-                $result.StandardOutput | Should -BeExactly 'Standard Output Message 0'
+                $result.StandardOutput | Should -BeExactly 'Standard Output Message'
 
-                $result.StandardError | Should -BeExactly 'Standard Error Message 0'
+                $result.StandardError | Should -BeExactly 'Standard Error Message'
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 0 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should not throw, return result with -PassThru, call Out-GitResult via -Verbose' {
+                $result = Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -PassThru -Verbose
+
+                $result.ExitCode | Should -BeExactly 0
+
+                $result.StandardOutput | Should -BeExactly 'Standard Output Message'
+
+                $result.StandardError | Should -BeExactly 'Standard Error Message'
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should not throw, return result with -PassThru, call Out-GitResult via -Debug' {
+                $result = Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -PassThru -Debug
+
+                $result.ExitCode | Should -BeExactly 0
+
+                $result.StandardOutput | Should -BeExactly 'Standard Output Message'
+
+                $result.StandardError | Should -BeExactly 'Standard Error Message'
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should not throw without -PassThru' {
+                { Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) } | Should -Not -Throw
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 0 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should not throw without -PassThru, call Out-GitResult via -Verbose' {
+                { Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -Verbose } | Should -Not -Throw
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should not throw without -PassThru, call Out-GitResult via -Debug' {
+                { Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -Debug } | Should -Not -Throw
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
 
                 Assert-VerifiableMock
             }
         }
 
-        Context 'When calling Invoke-Git with an access token' {
-            BeforeAll {
-                $mockProcess | Add-Member -MemberType ScriptProperty -Name ExitCode -Value { 1 } -Force
-
-                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardOutput' -Value {
-                    New-Object -TypeName 'Object' | `
-                        Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Output Message 1' } -PassThru -Force
-                } -Force
-
-                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardError' -Value {
-                    New-Object -TypeName 'Object' | `
-                        Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Error Message 1' } -PassThru -Force
-                } -Force
-
-                Mock -CommandName Write-Debug
-            }
-
-            It 'Should complete with ExitCode=1 and mask access token in debug message' {
-
-                $result = Invoke-Git -WorkingDirectory $TestDrive `
-                            -Arguments @( 'remote', 'set-url', 'origin', 'https://name:5ea239f132736de237492ff3@github.com/repository.wiki.git' ) `
-                            -Debug
-
-                $result.ExitCode | Should -BeExactly 1
-
-                $result.StandardOutput | Should -BeExactly 'Standard Output Message 1'
-
-                $result.StandardError | Should -BeExactly 'Standard Error Message 1'
-
-                Assert-MockCalled -CommandName Write-Debug -ParameterFilter {
-                    $Message -match 'https://name:RedactedToken@github.com/repository.wiki.git'
-                } -Exactly -Times 1 -Scope It
-
-                Assert-VerifiableMock
-            }
-        }
-
-        Context 'When git exits with an error code higher than 1' {
+        Context 'When git ExitCode -ne 0' {
             BeforeAll {
                 $mockProcess | Add-Member -MemberType ScriptProperty -Name 'ExitCode' -Value { 128 } -Force
-
-                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardOutput' -Value {
-                    New-Object -TypeName 'Object' | `
-                        Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Output Message 128' } -PassThru -Force
-                } -Force
-
-                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'StandardError' -Value {
-                    New-Object -TypeName 'Object' | `
-                        Add-Member -MemberType ScriptMethod -Name 'ReadToEnd' -Value { 'Standard Error Message 128' } -PassThru -Force
-                } -Force
             }
 
-            It 'Should complete with ExitCode=128' {
-                $result = Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status' )
+            It 'Should not throw, return result with -PassThru' {
+                $result = Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -PassThru
 
                 $result.ExitCode | Should -BeExactly 128
 
-                $result.StandardOutput | Should -BeExactly 'Standard Output Message 128'
+                $result.StandardOutput | Should -BeExactly 'Standard Output Message'
 
-                $result.StandardError | Should -BeExactly 'Standard Error Message 128'
+                $result.StandardError | Should -BeExactly 'Standard Error Message'
 
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 0 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should not throw, return result with -PassThru, call Out-GitResult via -Verbose' {
+                $result = Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -PassThru -Verbose
+
+                $result.ExitCode | Should -BeExactly 128
+
+                $result.StandardOutput | Should -BeExactly 'Standard Output Message'
+
+                $result.StandardError | Should -BeExactly 'Standard Error Message'
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should not throw, return result with -PassThru, call Out-GitResult via -Debug' {
+                $result = Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -PassThru -Debug
+
+                $result.ExitCode | Should -BeExactly 128
+
+                $result.StandardOutput | Should -BeExactly 'Standard Output Message'
+
+                $result.StandardError | Should -BeExactly 'Standard Error Message'
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should throw without -PassThru' {
+                { Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) } | Should -Throw
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 0 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should throw without -PassThru, call Out-GitResult via -Verbose' {
+                { Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -Verbose } | Should -Throw
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
+
+                Assert-VerifiableMock
+            }
+
+            It 'Should throw without -PassThru, call Out-GitResult via -Debug' {
+                { Invoke-Git -WorkingDirectory $TestDrive -Arguments @( 'status', '--verbose' ) -Debug } | Should -Throw
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 1 -Scope It
+
+                Assert-VerifiableMock
+            }
+        }
+
+        Context 'When throwing an error' {
+            BeforeAll {
+                $mockProcess | Add-Member -MemberType ScriptProperty -Name 'ExitCode' -Value { 128 } -Force
+
+                $tokenPrefix = ('pousr').ToCharArray() | Get-Random
+                $newTokenLength = Get-Random -Minimum 1 -Maximum 251
+                $newToken = (1..$newTokenLength | %{ ('abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890').ToCharArray() | Get-Random }) -join ''
+            }
+
+            $testCases = @(
+                @{
+                    'Command'      = @('status','--verbose');
+                    'ErrorMessage' = 'status --verbose';
+                },
+                @{
+                    'Command'      = @( 'remote','add','origin',"https://gh$($tokenPrefix)_$($newToken)@github.com/owner/repo.git" );
+                    'ErrorMessage' = 'remote add origin https://**REDACTED-TOKEN**@github.com/owner/repo.git';
+                }
+            )
+
+            It "Should throw exact with '<ErrorMessage>'" -TestCases $testCases {
+                param( $Command, $ErrorMessage )
+
+                $throwMessage = "$($script:localizedData.InvokeGitCommandDebug -f $ErrorMessage)`n" +`
+                                "$($script:localizedData.InvokeGitExitCodeMessage -f 128)`n" +`
+                                "$($script:localizedData.InvokeGitStandardOutputMessage -f 'Standard Output Message')`n" +`
+                                "$($script:localizedData.InvokeGitStandardErrorMessage -f 'Standard Error Message')`n"
+
+                { Invoke-Git -WorkingDirectory $TestDrive -Arguments $Command } | Should -Throw $throwMessage
+
+                Assert-MockCalled -CommandName Out-GitResult -Exactly -Times 0 -Scope It
                 Assert-VerifiableMock
             }
         }